+++ safe/safe/AppConfig.swift
@@ -12,7 +12,6 @@ struct AppConfig: Codable {
     var encryptionServiceChainId: Int
     var nodeServiceConfig: NodeServiceConfig
     var relayServiceURL: URL
-    var transactionHistoryServiceURL: URL
     var notificationServiceURL: URL
     var transactionWebURLFormat: String
     var chromeExtensionURL: URL
@@ -32,8 +31,7 @@ struct AppConfig: Codable {
     enum CodingKeys: String, CodingKey {
         case encryptionServiceChainId = "encryption_service_chain_id"
         case nodeServiceConfig = "node_service"
-        case relayServiceURL = "relay_service_url"
-        case transactionHistoryServiceURL = "transaction_history_service_url"
+        case relayServiceURL = "relay_service_url"        
         case notificationServiceURL = "notification_service_url"
         case transactionWebURLFormat = "transaction_web_url_format"
         case chromeExtensionURL = "chrome_extension_url"
